LayoutTests/platform/ios-sim-deprecated/http/tests/local/formdata/send-form-data-with-filename-expected.txt

CONSOLE MESSAGE: line 59: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(filesToDrag)')

Test that filename passed to FormData.append() takes precedence over filename in File.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Sending FormData containing one file with custom filename:
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
